server/server_job.py: remove legacy method of redirecting output
This also has the effect of losing the separate parser log file,
but, since the parser now uses the standard logging interface,
all messages will show up in the main log.

Note: It was discussed that it would be better to keep the original
parser log file, and I actually agreed about it. But, doing a little
more investigation, it turned out to be better to *not* do it.

Reasons include the fact that the redirect was only done because
using the old logging mechanism (print) the log messages would be
lost if not redirected to a file. Also, the parser log file is not
structured in any form, and not consumed by any other tool. Finally,
the location and existence of the parser log file was so obscure
that we did not know about it.

CC: Lucas Meneghel Rodrigues <lmr@redhat.com>
Signed-off-by: Cleber Rosa <crosa@redhat.com>