client.shared: Introduce a reporting library
Introduce a library specialized in client job reporting
that can:

 * Parse the autotest job 'status' file, using the infra
structure present on client.shared.base_job and create
a json data structure that will be dumped at the same
level of the status file
 * Generate and write an HTML report, pretty similar in
design and structure to the one currently produced by
client/tools/html_report.py.

It sports a more structured and clean approach at
creating these reports, which will help the maintainability
of the code base, and it also opens possibilities, such
as also creating HTML scheduler emails, as well as
output to other formats.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>