client.shared: Introduce jsontemplate
Our plan is to create a machine consumable file with
relevant summary data of an autotest client job. JSON
is the format of choice for many applications, due to
its simplicity and human readability.

jsontemplate is a simple templating engine that can
help render a nicely, human readable report of the
job data by using the JSON file. Its author, Andy
Chu, graciously agreed with double licensing this
file under both the Apache license version 2 (original
license) and the GPL version 2 (autotest's license).
The reason why it has to be done is that both licenses
have incompatibilities.

CC: Andy Chu <andychup@gmail.com>
Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>