backports: Replace json with simplejson
The backported json module from py 2.7 was still giving
us trouble when running on python 2.4. In order to fix
this problem, it is better to use a known and tested python
2.4 functional implementation rather than trying to hack our
own. So pick simplejson 2.1.0:

https://github.com/simplejson/simplejson/archive/v2.1.0.tar.gz

And replace the backport with this version. This solves
the issues we've been seeing under python 2.4 systems.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>