client/shared/distro_def.py: added new module
This module defines a structure and portable format for storing
relevant information on Linux Distributions.

The main goal is to be able to create a knowledge base of Linux
Distributions, either to be added to an official archive, or
created by local users.

This information will be used by the component isolation feature
to record differences on the systems running tests against the
distro official characteristics.

Changes from v1:
   * whitespace cleanup

Signed-off-by: Cleber Rosa <crosa@redhat.com>