CODING_STYLE: Update document reflecting the move to PEP8
While our base coding style was already PEP8, we had some
differences, such as 2 lines separating functions, and
single line imports. This made hard to automatically
enforce coding style rules other than review.

Given the incentives to move closer to the PEP8, we
decided to go forward with the plan, so the coding
style document has to be updated accordingly.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>