scheduler: add a option for how drone machines are deployed
That is, if they're supposed to be already properly configured
or if the scheduler should try to setup their dependencies.

Since most deployment scenarios of Autotest server/scheduler
involves only a single server and a single drone (localhost),
let's make the default 'True', that is, the machine is already
setup.

In more complex deployment scenarios, with multiple drones,
it would simply be matter of flipping this switch right along
with adding the additional drone hostnames.

Signed-off-by: Cleber Rosa <crosa@redhat.com>