global_config: add utility function/script to change global_config values
This keeps using the lowest-level (dumbest) approach of reading all
lines in the configuration file and changing the line that matches the
section and key provided.

The idea is to use this script during the build phase of packages
and avoid having a patch to change configuration files. This approach
is already partially used on the Fedora package but lacks consistency.

Signed-off-by: Cleber Rosa <crosa@redhat.com>