RPM packages: add SPEC file
The RPM spec file is the recipe for building RPMs. This one is
a "fork" of the Fedora RPM and tracks the state of the next
branch in autotest. That is, it's intended to be kept in sync
with the autotest code so that we can have quick and fresh
packages for each release.

Signed-off-by: Cleber Rosa <crosa@redhat.com>