install_autotest_server: RFC convert install script to python
WARNING: NOT FULLY WORKING.

This is an unfinished conversion of the install script to
python.  I started this partly for fun and partly because I was looking
at making modifications to the installer, including changing the git
cloning to fit our internal repo structure.  I was also hoping to
parallelize some of the install to speed it up where possible to speed
up install regression.

This includes a bunch of extra code for Threads and other Async helper
functions.

Signed-off-by: Ross Brattain <ross.b.brattain@intel.com>