database manager: use values that are known defaults for MySQL
I had the feeling that passing None to some connection parameters in
MySQLdb would make it use MySQL's default values. Looks like either
I was wrong from the begining, or new versions of MySQLdb changed
how they deal with None values.

So, for the hostname, default to 'localhost', for the admin user
default to 'root' and empty password.

Signed-off-by: Cleber Rosa <crosa@redhat.com>