frontend/setup_test_environment.py: use a test MySQL database
For running unittests. This will now DROP by default an existing
database if one exists with name "autotest_web_unittest_run".

Since it's highly unlikely that a user will host a MySQL database
with that name for other purposes, we consider it's safe enough
to drop it before each database related unittests.

Signed-off-by: Cleber Rosa <crosa@redhat.com>