scheduler unittests: use a valid metahost specification
There was a false assigment of metahost ids using 0, which looks like was
accepted by django+sqlite as valid data. That's not the case when running
the unittest on a MySQL database.

Since many labels are created on test setUp(), we can rely on using
metahost = 1 (the first label created).

Signed-off-by: Cleber Rosa <crosa@redhat.com>