utils.external_packages: Lower minimum requirements for psutil
This is a fix for issue #788.

Looking at the psutils documentation:

https://code.google.com/p/psutil/wiki/Documentation#System_related_functions

psutil.process_iter() is available in 0.4, it only had
a behavior change in 0.6 that should be largely irrelevant
to the autotest code. So, in order to help people running
older distros, such as Fedora 17, let's lower the minimum
version requirement for psutil.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>