Extend the lv_utils with full ramdisk cleanup and mount/umount functionality.
The ramdisk that the lv_utils module was using to create a volume group on
top of the RAM wasn't removed completely filling up the swap memory if
repeated multiple times. The only way to completely clean the space reserved
for no longer used volume group was to reset the machine which is unclean
way of doing this. The following patch adds a complete cleanup solving
these problems. In addition, it adds two small functions for mounting and
unmounting the logical volumes.

Signed-off-by: Plamen Dimitrov <pdimitrov@pevogam.com>