Parse distinct results from duplicating tests.
Very often the parsed results originate from different level
status files where the upper level status already contains
the test results from the lower lovel one. It is possible that
the upper level status contains extra test results, therefore
it cannot simply be ignored and must be parsed as well. In
order to allow this, we need to parse the distinct test results
and avoid flooding the console report with redundant such.

Signed-off-by: Plamen Dimitrov <pdimitrov@pevogam.com>