documentation: Workaround for genererating API docs
Put a symlink pointing the root of the git checkout
to 'autotest', to guarantee we'll have an 'autotest'
as the root directory, and we can generate the docs.

Fixing it properly will involve separate the autotest
api to a sub directory called 'autotest', and it will
take some time to finish.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>