Fix git issue with branch different than master.
When setting to fetch from a remote branch different than master, the
local repository is not enable to checkout to the branch due to hardcode
name 'master' used to setup the local branch name.

This patch changes the default name of lbranch variables from 'master'
to be None, and then setup in code the value to be the same of remote
branch name, usually specified by user and as already explained in
documentation.

Signed-off-by: Paulo Vital <paulo.vital@profitbricks.com>