install_server: if install times out, disable netboot
Currently, if an install times out, the host is left in the
to-be-installed state. However, this can result (if many hosts fail to
install) in all hosts being setup for netboot without much indication.
And unexpected user conditions also occur if someone reboots that system
to use it's prior existing OS. It is better to leave the system in
as-close to an unaltered state as possible.

Signed-off-by: Nishanth Aravamudan <nacc@linux.vnet.ibm.com>