pylint: Fix unsubscriptable values
The abstract classes initialized values as None and mentioned that
inherited classes must overwrite the values. This causes E1136 warning
as None is not subscriptable. I checked the value is not compared to
None anywhere (within the module) so it should be safe to put a valid
default and avoid this error.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>