autotest: The Big Import Cleanup
For a long time we had the 'common' imports as the big
problem when thinking about getting rid of unused imports
in the codebase. Now. with the effort of tagging the
common uses (along with other libraries that silently
do something when imported), it's possible to finally
reorganize the imports in the project all at once.

A lot of care has been taken, and testing has followed,
but there might be some problems. Keep in mind this is
done using a best-effort approach.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@scylladb.com>