client.shared.settings: Cope with standard python config locations
In linux world the configs are most commonly stored in "/etc", but in
purely python world they are "resources" as every other asset. Let's
change the way we discover the locations of those file to respect the
python defaults while still prioritizing the "/etc" as was before for
compatibility reasons.

This should suit most cases as distribution packages (RPM) can put the
config files in /etc but "python setup.py install" will use the
"resource" path (usually /usr/lib/python2.7/...) while still respecting
the /etc if created manually by users.

This is needed to avoid creating files in "/etc", which is not allowed
in easy_install or in non-priviledged mode.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>